+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -20,6 +20,8 @@
 ## Upcoming Preview Release
 * Introduced Restore-AzCosmosDBGremlinDatabase, Restore-AzCosmosDBGremlinGraph to restore deleted database and graph in the same account for Gremlin.
 * Introduced Restore-AzCosmosDBTable to restore deleted table in the same account.
+* New-AzCosmosDBAccount cmdlet was updated with new paramter: EnableBurstCapacity.
+* Update-AzCosmosDBAccount cmdlet was updated with: EnableBurstCapacity.
 
 ## Upcoming Release
 * Added PublicNetworkAccess parameter to `Restore-AzCosmosDBAccount`.