+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 
 ## Upcoming Release
+
+## Version 1.10.1
 * Updated Azure.Core to 1.31.0.
 
 ## Version 1.10.0